st40reader now does reading
